feat(server): add Sequelize setup and authentication logic  Added the Sequelize constructor, required the dotenv package to set environment variables, and included an if statement for the app to use the database when deployed or a local database otherwise.   Added authentication to check if the user is not logged in and redirect to the login route; if the user is logged in, execute the route function to allow them to view the page.  Included the necessary modules, created the app variable to create the Express app, defined the port to connect to, defined the session object, added middleware, and synced the Sequelize models to the server when it starts.
feat(server): add Sequelize setup and authentication logic

Added the Sequelize constructor, required the dotenv package to set environment variables, and included an if statement for the app to use the database when deployed or a local database otherwise.

Added authentication to check if the user is not logged in and redirect to the login route; if the user is logged in, execute the route function to allow them to view the page.

Included the necessary modules, created the app variable to create the Express app, defined the port to connect to, defined the session object, added middleware, and synced the Sequelize models to the server when it starts.