refactor(project): initiate full file restructure and add key functionalities  Initiated a full file restructure, established seeding file, added layouts, added middleware to the api folder and to the root of the controller folder. Added functionality for logging in and out as well as a signup option when the user is logged out. Added comments for clarity to the files so we know what we are doing with them.  BREAKING CHANGE: Comprehensive project restructure with new middleware, authentication features, and detailed comments.
refactor(project): initiate full file restructure and add key functionalities

Initiated a full file restructure, established seeding file, added layouts, added middleware to the api folder and to the root of the controller folder. Added functionality for logging in and out as well as a signup option when the user is logged out. Added comments for clarity to the files so we know what we are doing with them.

BREAKING CHANGE: Comprehensive project restructure with new middleware, authentication features, and detailed comments.