fix(media): fix media screen and polish completely  Adjusted the layout and styling to ensure that the media screen works correctly across different devices. Improved the alignment and appearance of the navigation bar and other elements to ensure a consistent look and feel. Ensured that the logo and title do not shift when the logout button appears.
fix(media): fix media screen and polish completely

Adjusted the layout and styling to ensure that the media screen works correctly across different devices. Improved the alignment and appearance of the navigation bar and other elements to ensure a consistent look and feel. Ensured that the logo and title do not shift when the logout button appears.